refactor(shared): チャットのスクロール追従とツール文言を shared へ移す
web だけが持っていた useStickToBottom と tool-display-text を shared に移し、
Thread 内に private 定義していたスクロールボタンを ScrollToBottomButton として
切り出した。widget から同じ挙動・同じ文言を使えるようにするための準備。
