+++ shared/src/lib/tool-display-text.ts
