+++ web/src/app/dashboard/components/mayor/MayorChatPanel.tsx
@@ -2,10 +2,10 @@ import { useChat } from "@ai-sdk/react";
 import { PaperAirplaneIcon, XMarkIcon } from "@heroicons/react/24/outline";
 import { MiniChatHeader } from "@nepp-chan/shared/components/MiniChatHeader";
 import { SpeechBubble } from "@nepp-chan/shared/components/SpeechBubble";
+import { useStickToBottom } from "@nepp-chan/shared/hooks/useStickToBottom";
 import { cn } from "@nepp-chan/shared/lib/class-merge";
 import { LoadingDots } from "@nepp-chan/shared/ui/Loading";
 import { useEffect, useMemo, useRef, useState } from "react";
-import { useStickToBottom } from "~/components/chat/hooks/useStickToBottom";
 import { MessageParts } from "~/components/chat/MessageParts";
 import { createThreadChatTransport } from "~/lib/api/chat-transport";
 import { threadRepository } from "~/lib/api/repository";
